Updated project organization to match montage one.
This involved moving run-test.html into test/run.html and moving
/test/all.js to /test/run.js. testacular.conf.js was removed because it
will not be used any more, and /test/data was removed because it will
not be used, at least not for initial tests. It can always be found in
the git history if it becomes needed.

/overview.html and /overview were removed because an overview is not
planned at this time.

package.json was updated to reflect the current state of this package.

/.gitignore, /LICENSE.md, and /README.md were updated with current
information.

The source in /data was not changed with this commit. Changed code
will be commited when the tests corresponding to the changed code
are added to /test/run.js.