Add number display layer & rename some components
This commit moves the display logic for numbers to a number-layer component. It
also renames display.reel to map.reel nad leaflet-map.reel to
leaflet-layer.reel. Finally, it moves the import of the Leaflet JavaScript and
CSS from leaflet-layer.html back to index.html to work around an apparent bug in
Montage where scripts imported in component HTML are not always loaded at the
right time when a page is reloaded.