Update copyright owner and trim package
This commit adjusts the copyright message in the LICENSE.md file to
refer to the new "Montage Studio" company name. Previously the copyright
message referred to the individual developers working on this project.

It also ajusts the package definition so the "data" folder, which
contains now obsolete code, is not included as part of the package.